refactor: rename core's core to controller
The module core within the core agent has been a pain for quite sometime now.
Renamed to controller, which I think makes sense since it's where the reconciler and
the scheduling logic lives.
