+++ control-plane/agents/core/src/controller/reconciler/nexus/garbage_collector.rs
@@ -1,4 +1,4 @@
-use crate::core::{
+use crate::controller::{
     reconciler::{GarbageCollect, PollContext, TaskPoller},
     specs::{OperationSequenceGuard, SpecOperations},
     task_poller::{PollEvent, PollResult, PollTimer, PollTriggerEvent, PollerState},
