+++ control-plane/agents/core/src/controller/reconciler/pool/mod.rs
@@ -1,4 +1,4 @@
-use crate::core::{
+use crate::controller::{
     reconciler::{GarbageCollect, ReCreate},
     specs::{OperationSequenceGuard, SpecOperations},
     task_poller::{PollContext, PollPeriods, PollResult, PollTimer, PollerState, TaskPoller},
