+++ control-plane/agents/core/src/controller/reconciler/volume/hot_spare.rs
@@ -1,4 +1,4 @@
-use crate::core::{
+use crate::controller::{
     reconciler::{nexus, PollContext, TaskPoller},
     specs::OperationSequenceGuard,
     task_poller::{squash_results, PollResult, PollerState},
