+++ control-plane/agents/core/src/controller/scheduling/resources/mod.rs
@@ -1,4 +1,4 @@
-use crate::core::{
+use crate::controller::{
     registry::Registry,
     wrapper::{NodeWrapper, PoolWrapper},
 };
