+++ control-plane/agents/core/src/node/registry.rs
@@ -1,4 +1,4 @@
-use crate::core::{registry::Registry, wrapper::NodeWrapper};
+use crate::controller::{registry::Registry, wrapper::NodeWrapper};
 use common::errors::SvcError;
 use common_lib::types::v0::transport::{NodeId, NodeState, Register};
