+++ control-plane/agents/core/src/pool/registry.rs
@@ -1,4 +1,4 @@
-use crate::core::{
+use crate::controller::{
     registry::Registry,
     wrapper::{GetterOps, *},
 };
