+++ control-plane/agents/core/src/server.rs
@@ -1,4 +1,4 @@
-pub mod core;
+pub mod controller;
 pub mod nexus;
 pub mod node;
 pub mod pool;
@@ -7,13 +7,12 @@ mod service;
 pub mod volume;
 pub mod watch;
 
+use controller::registry::NumRebuilds;
+use utils::{version_info_str, DEFAULT_GRPC_SERVER_ADDR};
+
 use http::Uri;
-
-use crate::core::registry::NumRebuilds;
-
 use opentelemetry::{trace::TracerProvider, KeyValue};
 use structopt::StructOpt;
-use utils::{version_info_str, DEFAULT_GRPC_SERVER_ADDR};
 
 #[derive(Debug, StructOpt)]
 #[structopt(name = utils::package_description!(), version = version_info_str!())]
@@ -96,7 +95,7 @@ async fn main() {
 
 async fn server(cli_args: CliArgs) {
     common_lib::init_cluster_info_or_panic().await;
-    let registry = core::registry::Registry::new(
+    let registry = controller::registry::Registry::new(
         cli_args.cache_period.into(),
         cli_args.store.clone(),
         cli_args.store_timeout.into(),
