+++ control-plane/agents/core/src/volume/scheduling.rs
@@ -1,4 +1,4 @@
-use crate::core::{
+use crate::controller::{
     registry::Registry,
     scheduling::{
         nexus,
