+++ control-plane/agents/core/src/volume/service.rs
@@ -1,4 +1,4 @@
-use crate::core::{
+use crate::controller::{
     registry::Registry,
     specs::{OperationSequenceGuard, ResourceSpecsLocked},
 };
